+++ MCG/MCG_CC/mcg_cc_supporter.py
@@ -5,7 +5,7 @@
 #       supporting methods and parameters reused by other classes.
 #
 #   COPYRIGHT:      Copyright (C) 2021-2022 Kamil Deć github.com/deckamil
-#   DATE:           7 JUL 2022
+#   DATE:           28 NOV 2022
 #
 #   LICENSE:
 #       This file is part of Mod Code Generator (MCG).
@@ -26,6 +26,9 @@
 #
 #       You should have received a copy of the GNU General Public License
 #       along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+
+from mcg_cc_error_handler import ErrorHandler
 
 
 # Description:
@@ -49,3 +52,69 @@ class Supporter(object):
     # name and uid, an example:
     # ADD fd5be3ed-0d38-42d0-ab56-d1058657eee8
     UID_OFFSET = -37
+
+    # This parameter defines start offset of name element after "name" marker in line of .exml file, i.e.
+    # number of characters after occurrence of "name" marker, where beginning of name element occurs, an example:
+    # <ID name="ADD" mc="Standard.OpaqueAction" uid="4f855500-ccdd-43a6-87d3-cc06dd16a59b"/>
+    NAME_START_OFFSET = 6
+
+    # This parameter defines end offset of name element before "mc" marker in line of .exml file, i.e.
+    # number of characters before occurrence of "mc" marker, where end of name element occurs, an example:
+    # <ID name="ADD" mc="Standard.OpaqueAction" uid="4f855500-ccdd-43a6-87d3-cc06dd16a59b"/>
+    NAME_END_OFFSET = -2
+
+    # This parameter defines start offset of uid element after "uid" marker in line of .exml file, i.e.
+    # number of characters after occurrence of "uid" marker, where beginning of uid element occurs, an example:
+    # <ID name="ADD" mc="Standard.OpaqueAction" uid="4f855500-ccdd-43a6-87d3-cc06dd16a59b"/>
+    UID_START_OFFSET = 5
+
+    # This parameter defines end offset of uid element before end of .exml file line, i.e.
+    # number of characters before occurrence of .exml file line end, where end of uid element occurs, an example:
+    # <ID name="ADD" mc="Standard.OpaqueAction" uid="4f855500-ccdd-43a6-87d3-cc06dd16a59b"/>
+    UID_END_OFFSET = -3
+
+    # Description:
+    # This method looks for <name> element within line of .exml file, an example of .exml file line:
+    # <ID name="ADD" mc="Standard.OpaqueAction" uid="4f855500-ccdd-43a6-87d3-cc06dd16a59b"/>
+    @staticmethod
+    def get_name(line, line_number):
+
+        # find position of name within the line
+        name_position = line.find("name")
+        # find position of mc within the line
+        mc_position = line.find("mc")
+
+        # check if <name> and <mc> position is found
+        if (name_position == -1) or (mc_position == -1):
+            # record error
+            ErrorHandler.record_error(ErrorHandler.GEN_ERR_NO_NAME_ELEMENT, line_number, "none")
+            # set error name
+            name = "NAME_NOT_FOUND"
+        else:
+            # get name
+            name = line[name_position + Supporter.NAME_START_OFFSET:mc_position + Supporter.NAME_END_OFFSET]
+
+        # return name
+        return name
+
+    # Description:
+    # This method looks for <uid> element within line of .exml file, an example of .exml file line:
+    # <ID name="ADD" mc="Standard.OpaqueAction" uid="4f855500-ccdd-43a6-87d3-cc06dd16a59b"/>
+    @staticmethod
+    def get_uid(line, line_number):
+
+        # find position of uid within the line
+        uid_position = line.find("uid")
+
+        # check if <uid> position is found
+        if uid_position == -1:
+            # record error
+            ErrorHandler.record_error(ErrorHandler.GEN_ERR_NO_UID_ELEMENT, line_number, "none")
+            # set error uid
+            uid = "UID_NOT_FOUND"
+        else:
+            # get uid
+            uid = line[uid_position + Supporter.UID_START_OFFSET:len(line) + Supporter.UID_END_OFFSET]
+
+        # return uid
+        return uid
